+++ includes/ObserverPattern.h
@@ -1,3 +1,6 @@
+#ifndef OBSERVER_PATTERN
+#define OBSERVER_PATTERN
+
 #include <vector>
 #include <string>
 using namespace std;
@@ -18,18 +21,18 @@ namespace ObserverPattern{
 	};	// End of class myObserverInterface
 
 	// Description:
-	// This abstract class is to send data to all observers. If you want
-	// use this class, you must define a new class to extand this class, 
-	// and then implement the abstract method `notifyData()`.
-	class mySubjectInterface{
+	// This class is to send data to all observers. If you want to use observer
+	// pattern, you must define a new class to extand this class for all 
+	// functions of this class.
+	class mySubject{
 		// protected fields
 		protected:
 			// The list is to store all observer.
 			vector<myObserverInterface*> m_vpoObserver;
 
 		public:
-			mySubjectInterface();
-			~mySubjectInterface();
+			mySubject();
+			~mySubject();
 
 			// Description: Register observer into the list
 			// @param poObserver: The object pointer of the class myObserverInterface
@@ -42,7 +45,8 @@ namespace ObserverPattern{
 			// Description: Notify the data to observers
 			// @param sKey: The keyword of the data
 			// @param pData: The data will notify to observer.
-			virtual void notifyData(const string sKey, void *pData) = 0;
-	};	// End of class mySubjectInterface
+			void notifyData(const string sKey, void *pData);
+	};	// End of class mySubject
 }	// End of namespace ObserverPattern
 
+#endif