+++ src/mySubject.cpp
@@ -1,16 +1,17 @@
 #include "../includes/ObserverPattern.h"
+#include <cstddef>
 
 using namespace ObserverPattern;
 
-mySubjectInterface::mySubjectInterface(){
+mySubject::mySubject(){
 	this->m_vpoObserver.clear();
 }	// End of constructor
 
-mySubjectInterface::~mySubjectInterface(){
+mySubject::~mySubject(){
 	this->m_vpoObserver.clear();
 }	// End of deconstructor
 
-void mySubjectInterface::registerObserver(myObserverInterface *poObserver){
+void mySubject::registerObserver(myObserverInterface *poObserver){
 	if(NULL == poObserver)
 		return;
 	
@@ -25,9 +26,9 @@ void mySubjectInterface::registerObserver(myObserverInterface *poObserver){
 	if(false == isExist){
 		this->m_vpoObserver.push_back(poObserver);
 	}	// End of if-condition
-}	// End of mySubjectInterface::registerObserver
+}	// End of mySubject::registerObserver
 
-void mySubjectInterface::removeObserver(myObserverInterface *poObserver){
+void mySubject::removeObserver(myObserverInterface *poObserver){
 	if(NULL == poObserver)
 		return;
 
@@ -46,4 +47,17 @@ void mySubjectInterface::removeObserver(myObserverInterface *poObserver){
 	}	// End of for-loop
 
 	this->m_vpoObserver.resize(iSizeOfObserverList - 1);
-}	// End of mySubjectInterface::removeObserver
+}	// End of mySubject::removeObserver
+
+void mySubject::notifyData(const string sKey, void *pData){
+	if(NULL == pData)
+		return;
+
+	size_t iSizeOfObserverList = this->m_vpoObserver.size();
+	if(iSizeOfObserverList <= 0)
+		return;
+
+	for(size_t i = 0; i < iSizeOfObserverList; i++){
+		this->m_vpoObserver[i]->updateData(sKey, pData);	
+	}	// End of for-loop
+}	// End of mySubject::notifyData